pyproject.toml
[tool.poetry]
name = "django-simple-blog"
version = "0.1.0"
description = ""
authors = ["Paw565pl <paw565pl@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
crispy-bootstrap5 = "^0.7"
django-crispy-forms = "^2.0"
pillow = "^10.0.0"
django-debug-toolbar = "^4.2.0"
django-resized = "^1.0.2"
psycopg = "^3.1.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
